move all image drawing stuff into separate file
